+++ ARMeilleure/State/FPSR.cs
@@ -3,7 +3,7 @@ using System;
 namespace ARMeilleure.State
 {
     [Flags]
-    public enum FPSR : ulong
+    public enum FPSR : uint
     {
         Ufc = 1 << 3,
         Qc  = 1 << 27,
